Update _config.yml
We have a text Logo and thus may not want to add an extra text title. Using a logo as the tile "title" works in a local Jekyll preview, but not when hosted online, causing an empty square at the top of the Menu. Trying to fix with this commit.